카드 컴포넌트, 컨텐츠(메뉴) 레이아웃 컴포넌트 분할
